1st attempt at moving the robot to its target.
Set the velocity exactly equal to the vector distance.  This means the
robot moves more quickly the further away it is and infinitesimally
slowly as it approaches the target (or w/in 10cm of the target).

I should add min & max velocity values.

I should tuck one or both of the arms before/while I move the robot
and/or do somthing about the fact that I hit the table with the arms
when I get close enough.
(Hint, run the following while the robot is moving:)

$ rosrun tf tf_echo /odom_combined /r_gripper_tool_frame

(This might give the position of the right hand gripper tool.  If it
does, then it shows that the gripper tool moves up a bit when we get
up to the table.)