Move forward until we're within 10cm of target, exit when we get there.
We used to (try to) move until we were w/in 1cm, and I never had the
patience to see if we ever got there.

Also, changed the target to 4.5m instead of 4m.  Now the arms bump
into the table.  I need to tuck them before I start moving.

Run:

$ rosrun tf tf_echo /odom_combined /r_gripper_tool_frame

to get an idea of where the right hand gripper tool is.