Use navigation stack to (attempt to) move robot to target location.
If you tuck the arms first with

$ rosrun pr2_tuckarm tuck_arms.py -r t -l t -q

and then run

$ rosrun wpdps1 mover2

and then wait a very long time, the robot will get very close to the
table, wiggle around a little bit, and then complain that it can't
reach it's goal.  I suspect there are some parameters about how close
to the table the robot is allowed to get that aren't set quite right.

I would like to add the ability to print out TF location messages
while the robot is moving.