Successfully picked up and placed 2 pieces... H2 to H3 G8 to H2
Added "pick up and place" loop to test5.py.

Modified piece model to have a wider base, with 1/2 the mass of the
piece so that the pieces would be more stable when placing them down.

I had tried to do this with computed <property> tags, but that didn't
seem to work.  So, for now (perhaps forever), I used global search and
replace to fill in the properties in chesspiece.urdf.macro.