Used the robot to move pieces around.
I looked at the torso lift joint in simulation at the start, it was at
0.0115.  I raised the torso fully to 0.300.  Then I ran test5 to the
point where it was about to close the gripper on a piece, set up the
table, chessboard, and a chess piece, and lowered the torso manually
to 0.273470237921 and watched while it moved pieces around.

It definitely knocks the piece to the right when lifting the arm to
move out of the way.  Also, the queen is about twice as tall as a
pawn, which I don't currently model.  But things more or less worked.